+++ tests/Microsoft.Bot.Builder.Tests/Skills/SkillConversationIdFactoryTests.cs
@@ -5,12 +5,11 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.Bot.Builder.Integration.Runtime.Skills;
 using Microsoft.Bot.Builder.Skills;
 using Microsoft.Bot.Schema;
 using Xunit;
 
-namespace Microsoft.Bot.Builder.Runtime.Tests
+namespace Microsoft.Bot.Builder.Tests.Skills
 {
     public class SkillConversationIdFactoryTests
     {
