+++ tests/Microsoft.Bot.Connector.Tests/Authentication/AllowedCallersClaimsValidationTests.cs
@@ -5,11 +5,10 @@ using System;
 using System.Collections.Generic;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using Microsoft.Bot.Builder.Integration.Runtime.Skills;
 using Microsoft.Bot.Connector.Authentication;
 using Xunit;
 
-namespace Microsoft.Bot.Builder.Runtime.Tests.Authentication
+namespace Microsoft.Bot.Connector.Tests.Authentication
 {
     public class AllowedCallersClaimsValidationTests
     {
